+++ packages/types/src/metadata/Metadata.spec.ts
@@ -1,9 +1,10 @@
 // Copyright 2017-2021 @polkadot/types authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import substrateData from '@polkadot/types-support/metadata/static-substrate';
+
 import { TypeRegistry } from '../create';
 import { Metadata } from './Metadata';
-import substrateData from './static-substrate';
 
 describe('Metadata', (): void => {
   it('allows creation from hex with JSON equivalence', (): void => {