+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Handle `BitSequence` in `PortableRegistry`
+- Move all test-support static files into `@polkadot/types-support`
+
+
 ## 5.3.2 Aug 3, 2021
 
 Upgrade priority: Low. Recommended for users of bridges.